UniqueParameter has a protected constructor to prevent manual construction of a UniqueParameter which bypasses the UniqueParameterNameGenerator
